+++ SensioFrameworkExtraBundle.php
@@ -1,25 +1,24 @@
 <?php
 
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Sensio\Bundle\FrameworkExtraBundle;
 
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Sensio\Bundle\FrameworkExtraBundle\DependencyInjection\Compiler\AddParamConverterPass;
-use Sensio\Bundle\FrameworkExtraBundle\DependencyInjection\Compiler\LegacyPass;
-
-/*
- * This file is part of the Symfony framework.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
 
 /**
  * SensioFrameworkExtraBundle.
  *
- * @author     Fabien Potencier <fabien@symfony.com>
+ * @author Fabien Potencier <fabien@symfony.com>
  */
 class SensioFrameworkExtraBundle extends Bundle
 {
@@ -28,6 +27,5 @@ class SensioFrameworkExtraBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new AddParamConverterPass());
-        $container->addCompilerPass(new LegacyPass());
     }
 }