+++ Templating/TemplateGuesser.php
@@ -14,20 +14,19 @@ namespace Sensio\Bundle\FrameworkExtraBundle\Templating;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Bundle\FrameworkBundle\Templating\TemplateReference;
 use Doctrine\Common\Util\ClassUtils;
 
 /**
  * The TemplateGuesser class handles the guessing of template name based on controller.
  *
- * @author     Fabien Potencier <fabien@symfony.com>
+ * @author Fabien Potencier <fabien@symfony.com>
  */
 class TemplateGuesser
 {
     /**
      * @var KernelInterface
      */
-    protected $kernel;
+    private $kernel;
 
     /**
      * @var string[]
@@ -35,10 +34,7 @@ class TemplateGuesser
     private $controllerPatterns;
 
     /**
-     * Constructor.
-     *
-     * @param KernelInterface $kernel             A KernelInterface instance
-     * @param string[]        $controllerPatterns Regexps extracting the controller name from its FQN.
+     * @param string[] $controllerPatterns Regexps extracting the controller name from its FQN
      */
     public function __construct(KernelInterface $kernel, array $controllerPatterns = array())
     {
@@ -53,14 +49,12 @@ class TemplateGuesser
      * and action names.
      *
      * @param callable $controller An array storing the controller object and action method
-     * @param Request  $request    A Request instance
-     * @param string   $engine
      *
-     * @return TemplateReference template reference
+     * @return string The template name
      *
      * @throws \InvalidArgumentException
      */
-    public function guessTemplateName($controller, Request $request, $engine = 'twig')
+    public function guessTemplateName($controller, Request $request)
     {
         if (is_object($controller) && method_exists($controller, '__invoke')) {
             $controller = array($controller, '__invoke');
@@ -88,9 +82,8 @@ class TemplateGuesser
             $matchAction = array(null, $controller[1]);
         }
 
-        $bundle = $this->getBundleForClass($className);
-
-        if ($bundle) {
+        $bundleName = null;
+        if ($bundle = $this->getBundleForClass($className)) {
             while ($bundleName = $bundle->getName()) {
                 if (null === $parentBundleName = $bundle->getParent()) {
                     $bundleName = $bundle->getName();
@@ -101,11 +94,11 @@ class TemplateGuesser
                 $bundles = $this->kernel->getBundle($parentBundleName, false);
                 $bundle = array_pop($bundles);
             }
-        } else {
-            $bundleName = null;
+
+            $bundleName = preg_replace('/Bundle$/', '', $bundleName);
         }
 
-        return new TemplateReference($bundleName, $matchController[1], $matchAction[1], $request->getRequestFormat(), $engine);
+        return sprintf(($bundleName ? '@'.$bundleName.'/' : '').$matchController[1].($matchController[1] ? '/' : '').$matchAction[1].'.'.$request->getRequestFormat().'.twig');
     }
 
     /**
@@ -115,7 +108,7 @@ class TemplateGuesser
      *
      * @return Bundle|null $bundle A Bundle instance
      */
-    protected function getBundleForClass($class)
+    private function getBundleForClass($class)
     {
         $reflectionClass = new \ReflectionClass($class);
         $bundles = $this->kernel->getBundles();