+++ Tests/Request/ParamConverter/ParamConverterManagerTest.php
@@ -168,8 +168,8 @@ class ParamConverterManagerTest extends \PHPUnit_Framework_TestCase
         $manager->apply($request, array($configuration));
     }
 
-    protected function createParamConverterMock()
+    private function createParamConverterMock()
     {
-        return $this->getMock('Sensio\Bundle\FrameworkExtraBundle\Request\ParamConverter\ParamConverterInterface');
+        return $this->getMockBuilder('Sensio\Bundle\FrameworkExtraBundle\Request\ParamConverter\ParamConverterInterface')->getMock();
     }
 }