+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 CHANGELOG
 =========
 
+4.0
+---
+
+ * removed support for the Templating component (only plain Twig is supported)
+ * removed unneeded phpdocs, converted protected to private properties
+ * bumped Symfony minimum version to 3.0
+ * bumped PHP minimum version to 5.5.9
+ * removed class parameters in container definitions
+
 3.0
 ---